devices_backend/requirements.txt
django
django-ninja
django-cors-headers
django-extensions
django-ninja-extra